#7 refactor : 코드 충돌 해결 및 타입 변경

+++ src/main/java/com/example/server/domain/content/repository/ContentDifficultyEvaluationRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface ContentDifficultyEvaluationRepository extends JpaRepository<ContentDifficultyEvaluation, Integer> {
-    Optional<ContentDifficultyEvaluation> findByUserIdAndContentId(int userId, int contentId);
+    Optional<ContentDifficultyEvaluation> findByUserIdAndContentId(Long userId, int contentId);
 }
